feat factories仮用意（シーダー生成できるように）

+++ backend/database/migrations/2022_04_20_075327_create_statistics_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId("questionnaire_id")->constrained()->comment("アンケートのid");
             $table->json("popular_words")->comment("人気の単語をjsonで収納");
-            $table->unsignedSmallInteger("number_of_responses")->comment("回答数");
+            $table->unsignedBigInteger("number_of_responses")->comment("回答数");
             $table->json("distribution_of_evaluations")->comment("評価の分散");
             $table->unsignedTinyInteger("emotion")->comment("感情数値化");
             $table->timestamps();
